requirements/common.txt
django==1.8
djangorestframework==3.1.1
mysql-python==1.2.5

boto==2.36.0
bleach==1.4.1
cssselect==0.9.1
django-countries==3.2
django-easy-select2==1.2.12
django-grappelli==2.6.4
django-hvad==1.2.0
django-imagekit==3.2.6
django-markdown-deux==1.0.5
-e git+https://github.com/eillarra/django-pagedown@c8a7b9acc6c66a44eeb9ceac91587bb8ad08ad89#egg=django_pagedown-master
django-storages==1.1.8
django-wkhtmltopdf==2.0.2
djrill==1.3.0
feedparser==5.1.3
lxml==3.4.2
pillow==2.7.0
python-magic==0.4.6
pytz==2015.2
requests==2.6.0
